TK-01038 - Correcciones en el sueldo de horas extra y en horas productivas

+++ trunk/code/self_managment/src/main/java/com/self_managment/service/TTSService.java
@@ -9,5 +9,5 @@ import com.self_managment.model.entity.TTS;
 public interface TTSService extends CRUDService<TTS, Serializable> {
 
 	public double getOvertimeSalary(Agent agent, int month, int year);
-	public double getProductiveHours(Agent agent, int month, int year);
+	public long getProductiveHours(Agent agent, int month, int year);
 }
